+++ src/stores/authStore.ts
@@ -81,9 +81,11 @@ export class AuthStore {
 
   sipShouldAuth = () =>
     this.sipState !== 'waiting' &&
-    !this.pbxLoginFromAnotherPlace &&
     this.sipState !== 'connecting' &&
     this.sipState !== 'success' &&
+    !this.pbxLoginFromAnotherPlace &&
+    !ctx.call.calls.length &&
+    !ctx.sip.phone?.getSessionCount() &&
     ((this.signedInId && this.sipPn.sipAuth) ||
       (this.pbxState === 'success' &&
         (this.sipState === 'stopped' ||
@@ -131,6 +133,7 @@ export class AuthStore {
   @observable ucConfig?: UcConfig
   @observable pbxConfig?: PbxGetProductInfoRes
   @observable listCustomPage: PbxCustomPage[] = []
+  @observable activeCustomPageId?: string
   saveActionOpenCustomPage = false
   customPageLoadings: { [k: string]: boolean } = {}
   getCustomPageById = (id: string) => this.listCustomPage.find(i => i.id == id)