+++ CHANGELOG.md
@@ -1,17 +1,22 @@
 #### 2.16.4
 
 - Initial implementation of [custom ringtone](./.doc/custom-ringtone.md)
+- Initial implementation of multiple custom pages
 - Add ctype=2 to pal url params, it should overwrite webphone.pal.param.ctype
 - Change error message to show "Internet connection failed" whenever an error occurred
 - Fix it should prevent button quickly double press
 - Fix it should handle deeplink while making call from PhoneAppli (issue 1014)
+- Fix it should have layout display well with status bar (issue 1020, 1049)
 - Fix it should save logs when the device is locked (issue 1040)
+- Fix ios loud speaker it should toggle well (1041)
 - Fix it should not display error message UC connection failure in incoming call (issue 1042)
 - Fix it should connect to sip after restart tomcat (issue 1043)
 - Fix it should not show notification with LPC service (issue 1044)
 - Fix it should not reload avatar when rerender (issue 1046)
 - Fix it should logout when app is closed from task manager (issue 1050)
 - Fix it should load video in video call (issue 1052)
+- Fix it should not hangup call when sip disconnect (issue 1056)
+- Fix it should clear keypad after hangup (issue 1057)
 
 #### 2.16.3